# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management & Garden Table API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority?: TaskInputPriority;
  dueDate?: string;
  tags?: string;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  priority?: TaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  tags?: string | null;
  completed?: boolean;
}

export type TaskStatsByPriority = {
  low: number;
  medium: number;
  high: number;
  urgent: number;
};

export interface TaskStats {
  total: number;
  completed: number;
  pending: number;
  overdue?: number;
  byPriority: TaskStatsByPriority;
}

export interface PriorityRequest {
  title: string;
  description?: string;
}

export type PrioritySuggestionPriority = typeof PrioritySuggestionPriority[keyof typeof PrioritySuggestionPriority];


export const PrioritySuggestionPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface PrioritySuggestion {
  priority: PrioritySuggestionPriority;
  reasoning: string;
  confidence?: number;
}

export interface GardenReservation {
  id: number;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  name: string;
  partySize: number;
  createdAt: string;
}

export interface GardenTokenInput {
  /** @minLength 1 */
  flat: string;
  /** @minLength 1 */
  token: string;
}

export interface GardenReservationInput {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  partySize: number;
}

export type ListTasksParams = {
status?: ListTasksStatus;
priority?: ListTasksPriority;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  all: 'all',
  pending: 'pending',
  completed: 'completed',
} as const;

export type ListTasksPriority = typeof ListTasksPriority[keyof typeof ListTasksPriority];


export const ListTasksPriority = {
  all: 'all',
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export type ListGardenReservationsParams = {
/**
 * ISO date string (YYYY-MM-DD) for the Monday of the week
 */
weekStart: string;
};