# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Management & Garden Table API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /garden/reservations:
    get:
      operationId: listGardenReservations
      tags: [garden]
      summary: List garden reservations for a week
      parameters:
        - name: weekStart
          in: query
          required: true
          schema:
            type: string
            description: ISO date string (YYYY-MM-DD) for the Monday of the week
      responses:
        "200":
          description: List of reservations for the week
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GardenReservation"
    post:
      operationId: createGardenReservation
      tags: [garden]
      summary: Reserve the garden table for an evening
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GardenReservationInput"
      responses:
        "201":
          description: Created reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenReservation"
        "409":
          description: Date already reserved
  /garden/tokens:
    post:
      operationId: registerGardenToken
      tags: [garden]
      summary: Register or update a device push token for a flat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GardenTokenInput"
      responses:
        "204":
          description: Token registered
  /garden/reservations/{id}:
    delete:
      operationId: deleteGardenReservation
      tags: [garden]
      summary: Cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    GardenReservation:
      type: object
      required: [id, date, name, partySize, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
          description: ISO date string (YYYY-MM-DD)
        name:
          type: string
        partySize:
          type: integer
        createdAt:
          type: string
    GardenTokenInput:
      type: object
      required: [flat, token]
      properties:
        flat:
          type: string
          minLength: 1
        token:
          type: string
          minLength: 1
    GardenReservationInput:
      type: object
      required: [date, name, partySize]
      properties:
        date:
          type: string
          description: ISO date string (YYYY-MM-DD)
        name:
          type: string
          minLength: 1
        partySize:
          type: integer
          minimum: 1
          maximum: 8