# lib/api-zod/src/generated/types/gardenTokenInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management & Garden Table API
 * OpenAPI spec version: 0.1.0
 */

export interface GardenTokenInput {
  /** @minLength 1 */
  flat: string;
  /** @minLength 1 */
  token: string;
}